testing a bare minimum of HITL functionality + added hitl installation for config.yml in .circleci
