+++ habitat-hitl/habitat_hitl/config/test_cfg.yaml
@@ -1,0 +1,20 @@
+# test_cfg.yaml
+
+# For more info about Hydra in Habitat, see habitat-lab/habitat/config/README.md
+# @package _global_
+
+defaults:
+  # We load the `pop_play` Habitat baseline featuring a Spot robot
+  # and a humanoid in HSSD scenes. See habitat-baselines/README.md.
+  - social_rearrange: pop_play
+  # Load default parameters for the HITL framework. See
+  # habitat-hitl/habitat_hitl/config/hitl_defaults.yaml.
+  - hitl_defaults
+  - _self_
+
+habitat_hitl:
+  window: ~
+  experimental:
+    headless:
+      do_headless: True
+      exit_after: 1
